Added a few PInvoke entries
Added a few PInvoke entries.

This is the real objective of this library, being able to craft custom
libraries from source and so avoid the "assembly hell" that can erupt
when consuming nuget assemblies and their dependencies.

To re-state, PInvoke is not the point, but just a convenient source of
useful code.